navigator.py
import asyncio
import logging
import os
from google import genai
from pydantic import BaseModel, Field

from memory import memory
from security import get_gemini_api_key

logger = logging.getLogger(__name__)

class ClickCoordinate(BaseModel):
    x: int
    y: int
    element_id: str
    action: str = Field(description="Action to perform: 'click' or 'type'")
    text_to_type: str = Field(default="", description="If action is type, the text to send")

class NavigatorAgent:
    """
    The 'Eyes' of Project Synapse.
    Uses Multimodal Vision to process legacy UI screens and output JSON actions.
    """
    def __init__(self, target_url: str):
        self.target_url = target_url
        self.api_key = get_gemini_api_key()
        
        # Initialize the fresh google-genai SDK 
        if self.api_key != "mock_gemini_key_for_testing":
            try:
                 self.client = genai.Client(api_key=self.api_key)
            except Exception as e:
                 logger.error(f"Failed to initialize GenAI client: {e}")
                 self.client = None
        else:
             self.client = None

    async def analyze_form_and_fill(self, merchant_data: dict) -> list[ClickCoordinate]:
        """
        Takes the known Merchant Data, 'looks' at the current URL (the legacy ERP),
        and returns exactly where to click and what to type.
        """
        logger.info(f"Navigator Agent navigating to {self.target_url}")
        
        # In a real scenario, we would use Playwright to capture a screenshot here.
        # For this hackathon demo, we simulate the Multimodal prompt generation.

        prompt = f"""
        You are a Multimodal UI Navigator Agent.
        Look at the provided screenshot of the legacy ERP dashboard.
        Your goal is to enter the following merchant data:
        Merchant Name: {merchant_data.get('name')}
        Tax ID: {merchant_data.get('tax_id')}
        Risk Score: {merchant_data.get('risk_score')}

        Identify the input fields and the submit button.
        Return ONLY valid JSON mapping these fields to their X/Y coordinates on the screen.
        """

        if self.client:
             logger.info("Calling Gemini 2.0 Flash Vision Multimodal API...")
             # Example of how we'd call the new SDK with structured output
             # response = self.client.models.generate_content(
             #     model='gemini-2.0-flash',
             #     contents=[screenshot_bytes, prompt],
             #     config=genai.types.GenerateContentConfig(
             #          response_mime_type="application/json",
             #          response_schema=list[ClickCoordinate]
             #     )
             # )
             pass
        
        # Simulated delay for model reasoning
        await asyncio.sleep(2)
        
        # Mock structured output reflecting what Gemini Flash 2.0 would return
        actions = [
            ClickCoordinate(x=150, y=200, element_id="merchantName", action="type", text_to_type=merchant_data.get("name", "")),
            ClickCoordinate(x=150, y=300, element_id="taxId", action="type", text_to_type=merchant_data.get("tax_id", "")),
            ClickCoordinate(x=150, y=400, element_id="riskScore", action="type", text_to_type=str(merchant_data.get("risk_score", 0))),
            ClickCoordinate(x=150, y=500, element_id="approvalStatus", action="type", text_to_type="APPROVED"),
            ClickCoordinate(x=400, y=600, element_id="btn-submit", action="click")
        ]
        
        logger.info(f"Navigator identified {len(actions)} required actions on the UI.")
        return actions

    async def execute_actions(self, actions: list[ClickCoordinate]):
        """
        Executes the generated coordinates using a browser automation tool (e.g. Playwright).
        """
        logger.info("Navigator Agent executing actions on legacy ERP...")
        for action in actions:
            if action.action == 'type':
                logger.info(f" -> Typing '{action.text_to_type}' into {action.element_id}")
            elif action.action == 'click':
                logger.info(f" -> Clicking button at ({action.x}, {action.y})")
            await asyncio.sleep(0.5)
        logger.info("All UI actions completed successfully.")
