requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-genai==0.3.0
websockets==12.0
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.18.0
redis==5.0.2
python-dotenv==1.0.1
pydantic==2.6.2
python-multipart==0.0.9
